Filter Flask requests by originating IP address.
This release represents the minimum viable product for restricting
access to a Heroku-hosted application to certain IP addresses.

This release trusts the X-Forwarded-For header Heroku adds to each
request. It is not safe in cases where the last IP address in this
header might be set by a malicious user.