Separate filter from rules
The IPFilter class now contains logic for interacting with Flask. The
Whitelist class contains logic for permitting hosts and networks and
checking whether an IP address should be allowed. This separation will
allow other types of rules to be created in the future.

This is actually a complete rewrite using a TDD approach.