Add callback option for filtering
The Callback class passes the source IP address to a callable defined by
the developer, which can return True to allow the connection or False to
block it. This is (slightly) simpler than making the developer define a
class that mimics Flask-IPFilter's Blacklist or Whitelist and provides a
path to customization that does not lock in the current structure of
pluggable filters.